% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAwa1.R
\name{CCAwa1}
\alias{CCAwa1}
\title{Constrained Correspondence Analysis.}
\usage{
CCAwa1(Y, X)
}
\arguments{
\item{Y}{Data matrix.}

\item{X}{Model matrix of constraints.}
}
\value{
\item{eig}{Eigenvalue of the first axis.} \item{w}{WA scores for
rows of the first axis.} \item{u}{LC scores for rows of the first axis.}
\item{v}{Column scores of the first axis.}
}
\description{
First axis of Constrained Correspondence Analysis using weighted averaging.
}
\details{
This is a classic algorithm to perform CCA, and is included to show how
simple it is. Constrained Correspondence Analysis was originally introduced
in ecology as a result of this algorithm (ter Braak 1986). Palmer (1993)
gives a very lucid explanation of the algorithm and its two kind of row
scores, known as Weighted Averages (WA) scores and Linear Combination (LC)
scores.  The function only finds one axis.  It would be possible to find
later axes by orthogonalizing against previous axes, but if several axes are
needed, it is better to use other algorithms from the beginning.
}
\references{
Palmer, M. W. (1993) Putting things in even better order: The
advantages of canonical correspondence analysis. \emph{Ecology} 74,
2215--2230.

ter Braak, C. J. F. (1986) Canonical correspondence analysis: a new
eigenvector technique of multivariate direct gradient analysis.
\emph{Ecology} 67, 1167--1179.
}
\seealso{
\code{\link{pCCA}} provides a better algorithm for
constrained correspondence analysis.
}
\author{
Jari Oksanen.
}
\keyword{multivariate}
