% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCA.R
\name{DCA}
\alias{DCA}
\title{Detrended Correspondence Analysis}
\usage{
DCA(Y, pairwise = FALSE, ...)
}
\arguments{
\item{Y}{Data matrix.}

\item{pairwise}{Detrend axis pairwise against each previous axis
similarly as in \code{\link[vegan]{decorana}}, or against all
previous axes simultaneously.}

\item{\dots}{Other arguments passed to \code{\link{loess}}.}
}
\value{
Function returns a list with items:
\item{DCA.eig}{Eigenvalue as estimated during detrending
    steps.}
\item{eig}{Eigenvalue as estimated by shrinking of
    scores in final solution, similarly as in \code{\link{CA}}.}
\item{u, v}{Row and column scores.}
}
\description{
Correspondence Analysis where axes are detrended by
\code{\link{loess}}.
}
\details{
Function performes detrended correspondence analysis but it differs
essentially from the classic \code{\link[vegan]{decorana}}. Not
only is the detrending method different, but \code{decorana} makes
much more than detrending. Most importantly it attempts to rescale
axes to constant response width (see \pkg{vegan} function
\code{\link[vegan]{tolerance}}) which often has a stronger effect
than detrending.

If you want to inspect the steps of \code{decorana}, function
\code{rdecorana} in github package \pkg{natto} implements the
function in plain \R{} which is much easier to follow than the
original Fortran code.
}
\seealso{
\code{\link[vegan]{decorana}}.
}
\author{
Jari Oksanen
}
\keyword{multivariate}
