% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCApot1.R
\name{PCApot1}
\alias{PCApot1}
\title{Principal Component Analysis}
\usage{
PCApot1(Y, scale = FALSE)
}
\arguments{
\item{Y}{Data matrix.}

\item{scale}{Scale columns to unit variance.}
}
\value{
The function returns an object with items:
    \item{eig}{Eigenvalue of the first axis.} \item{ueig}{Rowscores
    scaled by eigenvalues.}  \item{v}{Orthonormal column scores.}
}
\description{
First axis of Principal Component Analysis using power method.
}
\details{
This is the poorest algorithm of performing PCA, and is included only to
show how simple it is. The function only finds one axis. It would be
possible to find later axes by orthogonalizing against previous axes, but
that is not worthwhile with an algorithm as poor as this one.
}
\references{
Jongman, R.H., ter Braak, C.J.F & van Tongeren,
    O.F.R. (1987) Data analysis in community and landscape
    ecology. Wageningen.
}
\seealso{
\code{\link{PCA}} and \code{\link{PCAeig}} for better
    algorithms.
}
\author{
Jari Oksanen.
}
\keyword{multivariate}
