% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wPCoA.R
\name{wPCoA}
\alias{wPCoA}
\title{Weighted Principal Coordinate Analysis}
\usage{
wPCoA(D, w)
}
\arguments{
\item{D}{Dissimilarity or distance structure as in \code{\link{PCoA}}.}

\item{w}{Weights for observations.}
}
\value{
Function returns an object from \code{\link{eigen}}.
\item{values}{Eigenvalues} \item{vectors}{Orthonormal row scores.}
}
\description{
Function performs Principal Coordinate Analysis with row weights.
}
\author{
Jari Oksanen
}
\seealso{
\code{\link{PCoA}}, \code{\link[vegan]{wcmdscale}}.
}
\keyword{multivariate}

