% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA.R
\name{CA}
\alias{CA}
\title{Correspondence Analysis}
\usage{
CA(Y)
}
\arguments{
\item{Y}{Data Matrix}
}
\value{
Function returns an object from \code{\link{svd}} with items:
\item{d}{Singular values which are the square roots of eigenvalues.}
\item{u}{Orthonormal row scores.} \item{v}{Orthonormal column scores.}
}
\description{
Correspondence Analysis via singular value decomposition.
}
\note{
Function \code{\link[vegan]{cca}} (\pkg{vegan}) package uses similar
algorithm.
}
\author{
Jari Oksanen
}
\references{
Greenacre, M. J. (1984). Theory and applications of
correspondence analysis. Academic Press, London.
}
\seealso{
\code{\link{pCCA}}, \code{\link{svd}}, \code{\link[vegan]{cca}}.
}
\keyword{multivariate}

