% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAdist.R
\name{CAdist}
\alias{CAdist}
\title{Correspondence Analysis}
\usage{
CAdist(Y)
}
\arguments{
\item{Y}{Data matrix.}
}
\value{
The function returns an object from \code{\link{wPCoA}} which in
turn returns an \code{\link{eigen}} result amended with species scores.
\item{values }{Eigenvalues.} \item{vectors }{Weighted orthonormal row
scores.} \item{v}{Weighted orthonormal column scores.}
}
\description{
Correspondence analysis as a weighted Principal Coordinate Analysis of
Chi-squared distances.
}
\author{
Jari Oksanen
}
\keyword{multivariate}
