% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbRDA.R
\name{dbRDA}
\alias{dbRDA}
\title{Distance-based Redundancy Analysis}
\usage{
dbRDA(D, X = NULL, Z = NULL)
}
\arguments{
\item{D}{Dissimilarity structure of dependent data as in \code{\link{PCoA}}.}

\item{X}{Model matrix of constraints (can be missing).}

\item{Z}{Model matrix of conditions (can be missing).}
}
\value{
The function returns the same object as \code{\link{pRDA}}.
}
\description{
Distance-based Redundancy Analysis is a Redundancy Analysis on the
eigenvalue-scaled result of Principal Coordinates Analysis.
}
\note{
This function does not actually perform distance-based RDA,
    but a simple variant called \code{\link[vegan]{capscale}} in
    \pkg{vegan}. The function will be fixed ASAP.
}
\references{
Legendre, P. & Anderson, M.J. (1999). Canonical analysis of
principal coordinates: a useful method of constrained ordination for
ecology. \emph{Ecology} 84, 511--525.
}
\seealso{
\code{\link{PCoA}}, \code{\link{pRDA}},
\code{\link[vegan]{capscale}}.
}
\author{
Jari Oksanen
}
\keyword{multivariate}
