% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCCA.R
\name{pCCA}
\alias{pCCA}
\title{[Partial] [Constrained] Correspondence Analysis.}
\usage{
pCCA(Y, X = NULL, Z = NULL)
}
\arguments{
\item{Y}{Dependent data Matrix.}

\item{X}{Model matrix of constraints (can be missing).}

\item{Z}{Model matrix of conditions (can be missing).}
}
\value{
later\dots{}
}
\description{
[Partial] [Constrained] Correspondence Analysis via singular value and QR
decomposition.
}
\note{
Function \code{\link[vegan]{cca}} (\pkg{vegan} package) is similar.
}
\author{
Jari Oksanen
}
\seealso{
\code{\link{pRDA}}, \code{\link{svd}}, \code{\link{qr}},
\code{\link[vegan]{cca}}.
}
\keyword{multivariate}

