% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{RDA}
\alias{RDA}
\title{Redundancy Analysis}
\usage{
RDA(Y, X, scale = FALSE)
}
\arguments{
\item{Y}{Dependent data matrix.}

\item{X}{Model matrix of constraints.}

\item{scale}{Scaling of dependent data.}
}
\value{
later\dots{}
}
\description{
Redundancy Analysis via QR and singular value decompositions.
}
\note{
Function \code{\link{pRDA}} is a more versatile an complicated
variant, that also can do partial RDA or PCA as special cases.
}
\author{
Jari Oksanen
}
\seealso{
\code{\link{pRDA}}, \code{\link{svd}}, \code{\link{qr}},
\code{\link[vegan]{rda}}.
}
\keyword{multivariate}

