% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wRDA.R
\name{wRDA}
\alias{wRDA}
\title{Weighted [Partial] Redundancy Analysis and Principal Components Analysis}
\usage{
wRDA(Y, X = NULL, Z = NULL, scale = FALSE, w = NULL, cw = NULL)
}
\arguments{
\item{Y}{Dependent Data Matrix.}

\item{X}{Model matrix of constraints (can be missing).}

\item{Z}{Model matrix of constraints (can be missing).}

\item{scale}{Scale dependent matrix.}

\item{w}{Row weights.}

\item{cw}{Column weights.}
}
\value{
Function returns similar object as \code{\link{pRDA}}.
}
\description{
Weighted [Partial] Redundancy Analysis via QR and singular value
decomposition or Principal Components Analysis via singular value
decomposition.
}
\note{
Function \code{\link{pRDA}} is a similmar function for non-weighted
[partial] RDA.
}
\seealso{
\code{\link{pRDA}}, \code{\link{pCCA}}, \code{\link[vegan]{rda}}.
}
\author{
Jari Oksanen
}
\keyword{multivariate}
